yuv2rgb: fix integer overflow in fill_gv_table()
Fixes CID703747

+++ libswscale/yuv2rgb.c
@@ -680,7 +680,7 @@ static void fill_table(uint8_t* table[256 + 2*YUVRGB_TABLE_HEADROOM], const int 
     }
 }
 
-static void fill_gv_table(int table[256 + 2*YUVRGB_TABLE_HEADROOM], const int elemsize, const int inc)
+static void fill_gv_table(int table[256 + 2*YUVRGB_TABLE_HEADROOM], const int elemsize, const int64_t inc)
 {
     int i;
     int off    = -(inc >> 9);
